Fix marker parentheses
The ant simulator/compiler expects a sense marker syntax like this:

Sense Ahead 0 378 Marker 4

Haskell however made it:

Sense Ahead 0 378 (Marker 4)

Because it saw the nesting of the marker condition in the sense statement.

I implemented the showInstruction function to fix this problem. It could be made prettier if there is some way in Haskell to override the default Show behaviour, but that sounds too OO-like so probably impossible.